+++ tests/src/07-gettype-of-runtime-value.ts
@@ -5,14 +5,18 @@ import {
 } from "tst-reflect";
 
 test("getType(val) called with runtime value and it is not Type.Unknown", () => {
-	@reflect()
+	// @reflect() decorator required in some configurations
 	class A
 	{
+		constructor(public foo: string)
+		{
+		}
 	}
 	
-	const a: unknown = new A();
+	const a: unknown = new A("Lipsum");
 
 	expect(getType(a) instanceof Type).toBe(true);
 	expect(getType(a)).not.toBe(Type.Unknown);
 	expect(getType(a).name).toBe("A");
+	expect(getType(a)).toBe(getType<A>());
 });