+++ tests/src/08-native-types.ts
@@ -1,4 +1,7 @@
-import { Type } from "tst-reflect";
+import {
+	getType,
+	Type
+} from "tst-reflect";
 
 test("All native types are correct", () => {
 	expect(Type.Unknown.name).toBe("unknown");
@@ -10,4 +13,32 @@ test("All native types are correct", () => {
 	expect(Type.String.name).toBe("String");
 	expect(Type.Boolean.name).toBe("Boolean");
 	expect(Type.Date.name).toBe("Date");
+});
+
+test("Type of number is Type.Number", () => {
+	expect(getType(5)).toBe(Type.Number);
+});
+
+test("Type of string is Type.String", () => {
+	expect(getType("foo")).toBe(Type.String);
+});
+
+test("Type of boolean is Type.Boolean", () => {
+	expect(getType(true)).toBe(Type.Boolean);
+});
+
+test("Type of undefined is Type.Undefined", () => {
+	expect(getType(undefined)).toBe(Type.Undefined);
+});
+
+test("Type of null is Type.Null", () => {
+	expect(getType(null)).toBe(Type.Null);
+});
+
+test("Type of Date is Type.Date", () => {
+	expect(getType(new Date())).toBe(Type.Date);
+});
+
+test("Base type of object literal is Type.Object", () => {
+	expect(getType({}).baseType).toBe(Type.Object);
 });