+++ tests/src/09-reflect-decorator.ts
@@ -3,12 +3,30 @@ import {
 	Type
 } from "tst-reflect";
 
-test("@reflect decorator force type to appear in metadata", () => {
-	@reflect()
-	class A
-	{
-	}
+@reflect()
+class A
+{
+}
 
-	expect(Type.getTypes()).toHaveLength(1);
+/**
+ * @reflect
+ */
+function foo()
+{
+	return function(_: any) {}
+}
+
+@foo()
+class B
+{
+}
+
+test("@reflect decorator force type to appear in metadata", () => {
+	expect(Type.getTypes()).toHaveLength(2);
 	expect(Type.getTypes()[0].name).toBe("A");
+});
+
+test("Custom tagged decorator force type to appear in metadata", () => {
+	expect(Type.getTypes()).toHaveLength(2);
+	expect(Type.getTypes()[1].name).toBe("B");
 });