+++ transformer/src/getTypeCall.ts
@@ -30,7 +30,7 @@ const creatingTypes: Array<number> = [];
  * @param context
  * @param typeCtor
  */
-export function getTypeCall(type: ts.Type, symbol: ts.Symbol | undefined, context: Context, typeCtor?: ts.EntityName | ts.DeclarationName): GetTypeCall // TODO: Remove symbol parameter
+export function getTypeCall(type: ts.Type, symbol: ts.Symbol | undefined, context: Context, typeCtor?: ts.EntityName | ts.DeclarationName): GetTypeCall // TODO: Remove symbol parameter if possible
 {
 	const id: number | undefined = (type.aliasSymbol || type.symbol as any)?.["id"];
 	let typePropertiesObjectLiteral: ts.ObjectLiteralExpression | undefined = undefined;